Fix SSH config for Advanced users + Description
Using just one host over the .ssh/config is returning a "Password" query that is not the correct First Factor (possibly going to /etc/passwd instead). To correctly make the config work we need to add the details of the lander node too. This way it will ask for the First factor and then the Second one, on the first jump. Then it will ask for a combination of the first factor and the second one, on the second jump.