Improve description of reseting temproary password
Over "Setting up an account on Kupe", step-6 it was not very evident how the process would really work. I have tested myself this using a test user and decided to remove the image (not updated any more) and use code to describe the example (easily editable for the future).
For now, it's best to show user how this is really working (even if not yet neat).

I have also added some separation bars between chapters for better readability.